required table column changes

+++ src/main/java/com/apmosys/employeeportal/model/PoDepartmentMapping.java
@@ -25,7 +25,7 @@ public class PoDepartmentMapping {
 	 @Id
 		@GeneratedValue(strategy = GenerationType.IDENTITY)
 		private Long poDepartmentMapId;
-		private Integer poId;
+		private Long poId;
 		private Long deptId;
 		private boolean active;
